Use touchstone-tools system and process charts
Updated for touchstone-tools v0.4.2.

Touchstone-tools uses gnuplot instead of R or Julia for generating
charts.  The dependencies for having gnuplot seem to be significantly
light than R of Julia, and gnuplot seems to generate charts a bit
faster.

Saving tens of minutes off report generation from tests run on large
systems seems to outweigh any opinions on visual differences between the
charting engines available from R or Julia.

gnuplot has become a bit more feature rich then when we originally use
it for report generation back in 2002-ish.

Touchstone-tools lays out the files a bit differently that what has been
done so a number of adjustment were made in where charts were expected
to be.

Removed the option to specify the output directory for reports.
Simplifies things for the time being by assuming charts are in a known
relative path to the reports.